Test using two var for loop
Since it is iterating over a module that uses for_each. I think a module using count might work with the single temporary var syntax but when using for_each you get a map, not a list, tuple, or set, and so it is not compatible with the single temporary var syntax.